Use integer last_seen_time for fading memory

+++ game/world/game_map.py
@@ -94,7 +94,7 @@ class GameMap:
             (height, width), dtype=np.float32, order="C"
         )
         self.last_seen_time: np.ndarray = np.zeros(
-            (height, width), dtype=np.float32, order="C"
+            (height, width), dtype=np.int32, order="C"
         )
         # Perception layers reused across turns
         self.noise_map: np.ndarray = np.zeros(
@@ -148,7 +148,7 @@ class GameMap:
         return self.transparent[y, x]
 
     # --- Memory fade helper ---
-    def update_memory_fade(self, current_time: float) -> None:
+    def update_memory_fade(self, current_time: int) -> None:
         """Fade remembered tiles based on elapsed time."""
         update_memory_fade(
             current_time,
